+++ invoice_to_stock/models/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from . import invoice_stock
+from . import stock_picking_type
