+++ invoice_to_stock/views/invoice_stock_view.xml
@@ -7,18 +7,18 @@
             <field name="inherit_id" ref="account.view_move_form"/>
             <field name="arch" type="xml">
                 <xpath expr="//header" position="inside">
-                    <field name="x_from_sale" invisible="1" />
+                    <field name="x_from_sale" invisible="1"/>
                     <button name="action_create_stock_picking" string="Afectar Stock" type="object" class="oe_highlight"
                             attrs="{'invisible':['|','|','|',('x_from_sale','=',True), ('move_type', 'not in', ('out_refund','in_refund')),('state', 'in', ('draft','cancel')),('x_picking_count', '>=', 1)]}"/>
                 </xpath>
                 <xpath expr="//field[@name='invoice_date']" position="after">
                     <field name="x_invoice_picking_id" invisible="1"/>
-                    <field name="x_origin_move" invisible="1" />
+                    <field name="x_origin_move" invisible="1"/>
                     <field name="x_picking_type_id"
                            options="{'no_create': True}"
                            attrs="{'invisible': [('move_type','in',('entry','out_invoice','in_invoice'))], 'readonly':['|', ('x_picking_count', '>=', 1), ('x_origin_move', '=', 'reversal')]}"
-                           domain="(context.get('default_move_type') in ('in_invoice','out_refund') and [('code', '=', 'incoming')])
-                                    or (context.get('default_move_type') in ('out_invoice', 'in_refund') and [('code', '=', 'outgoing')])" />
+                           domain="(context.get('default_move_type') in ('in_invoice','out_refund') and ['&amp;', ('code', '=', 'incoming'), ('affects_stock', '=', 'True')])
+                                    or (context.get('default_move_type') in ('out_invoice', 'in_refund') and ['&amp;', ('code', '=', 'outgoing'), ('affects_stock', '=', 'True')])"/>
                 </xpath>
                 <xpath expr="//div[hasclass('oe_button_box')]" position="inside">
                     <button type="object"
@@ -32,5 +32,4 @@
             </field>
         </record>
     </data>
-
 </odoo>
