+++ invoice_to_stock/views/stock_picking_type_view.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data>
+        <record id="xinvoice_stock_picking_type_form_view" model="ir.ui.view">
+            <field name="name">xinvoice.stock.picking.type.form</field>
+            <field name="model">stock.picking.type</field>
+            <field name="inherit_id" ref="stock.view_picking_type_form"/>
+            <field name="arch" type="xml">
+                <xpath expr="//field[@name='show_reserved']" position="before">
+                    <field name="affects_stock"/>
+                </xpath>
+            </field>
+        </record>
+    </data>
+</odoo>
